Filtered reports by student ID in dashboard and reports screens, updated report data models for null safety, added error handling to report submission, and removed unused UserRepository.

+++ app/src/main/java/com/example/campuscare10/model/Report.kt
@@ -15,7 +15,7 @@ data class Report(
     @SerialName("submitted_by")
     val submittedBy: String?,
     val status: String,
-    val rating: Float? = null,
+    val rating: Double? = null,
     val note: String? = null,
     @SerialName("image_uri")
     val imageUri: String?
